refactor: 프로젝트 폴더 수정
