+++ CleanArea/CleanArea/iOS/Detail/DetailSummaryView.swift
@@ -88,35 +88,7 @@ struct DetailSummaryView: View {
     }
 }
 
-struct TagModifier: ViewModifier {
-    func body(content: Content) -> some View {
-        content
-            .bold()
-            .background(
-                Rectangle()
-                    .cornerRadius(10)
-                    .foregroundStyle(Color.init(hex: "E1F1E1"))
-                    .shadow(color: Color(uiColor: UIColor.systemGray3), radius: 5, x: 5, y: 5)
-            )
-    }
-}
 
-struct DetailSemiTitleModifier: ViewModifier {
-    func body(content: Content) -> some View {
-        content
-            .multilineTextAlignment(.leading)
-            .font(.system(size: 17, weight: .semibold))
-            .foregroundStyle(Color.init(hex: "111111"))
-    }
-}
-
-struct DetailSemiContentModifier: ViewModifier {
-    func body(content: Content) -> some View {
-        content
-            .multilineTextAlignment(.leading)
-            .font(.system(size: 17))
-    }
-}
 
 /*
 #Preview {
