+++ CleanArea/CleanArea/iOS/Model/RecommandCellModel.swift
